feat: address feature request #99239
This commit addresses #99239 as a proof of concept.  The filter list is
now constrained only to geo fields that are actually used within the
layer sources.  Because the ultimate filter created is independent of
the index pattern, the UI instead shows the field type and deduplicates
when the same field exist in multiple index pattern.

This commit leaves room to refactor out some unneeded types and does not
yet add unittests.